Move PaymentDataValidator out of todo/
Into the Payment BC since it deals with payment stuff and has
a "payment type", which while not a PHP type, is a concept from
this BC.
