+++ todo/Tests/Unit/Validation/PaymentDataValidatorTest.php
@@ -5,10 +5,10 @@ declare( strict_types = 1 );
 namespace WMDE\Fundraising\Frontend\Tests\Unit\Validation;
 
 use WMDE\Euro\Euro;
-use WMDE\Fundraising\Frontend\Validation\PaymentDataValidator;
+use WMDE\Fundraising\Frontend\PaymentContext\Domain\PaymentDataValidator;
 
 /**
- * @covers \WMDE\Fundraising\Frontend\Validation\PaymentDataValidator
+ * @covers \WMDE\Fundraising\Frontend\PaymentContext\Domain\PaymentDataValidator
  *
  * @licence GNU GPL v2+
  * @author Kai Nissen < kai.nissen@wikimedia.de >
